Manage time-chart context in component
Much of the time-chart state is now managed in its own component. It wraps an existing range-slider component, as well as a c3 component that is not built yet.

Timeline attributes, which are array of objects with 2 properties, status and month, are needed at various states of aggregation for this application, including: flattened, sorted, and composed with other models taht have Timelines. This introduces new computed properties that expose the data at various aggregate states.